requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytest==7.4.3
reportlab==4.0.7
PyMuPDF==1.23.8
groq==0.4.1
python-dotenv==1.0.0
cryptography==41.0.7
Werkzeug==3.0.1